Move HexOffset helper to the deduplication package as it's the only use
