+++ ledger/participant-state/kvutils/src/test/suite/scala/com/daml/ledger/participant/state/kvutils/deduplication/HexOffsetSpec.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.offset
+package com.daml.ledger.participant.state.kvutils.deduplication
 
 import com.daml.lf.data.Ref
 import org.scalatest.matchers.should.Matchers
